🌟feat: 안내 뷰 컴포넌트 파일 분리
